[Refactor] #196 - 프로필 컴바인 기반 코드 수정
- 뷰에서 디바운스 처리 후 뷰모델 요청
